requirements.txt
python-telegram-bot==21.4
requests==2.31.0
python-dotenv==1.0.0
pillow==10.2.0


